requirements.txt
pandas
numpy
google-cloud-storage
scikit-learn 
pyyaml 
ipykernel
matplotlib
seaborn 
statsmodels
imbalanced-learn
xgboost
lightgbm
mlflow